requirements.txt
bcrypt==3.1.4
blinker==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycparser==2.18
requests==2.18.4
sightengine==1.3.0
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
